added friendship and person api's

+++ socialapp/src/main/java/com/shivang/socialapp/service/impl/OrganizationServiceImpl.java
@@ -1,15 +1,22 @@
 package com.shivang.socialapp.service.impl;
 
+import org.hibernate.HibernateException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.shivang.socialapp.dao.OrganizationDAO;
+import com.shivang.socialapp.dao.PersonDAO;
 import com.shivang.socialapp.model.Organization;
 import com.shivang.socialapp.service.OrganizationService;
 
 @Service
+@Transactional
 public class OrganizationServiceImpl implements OrganizationService{
 
+	@Autowired
+	PersonDAO personDAO;
+	
 	@Autowired
 	OrganizationDAO organizationDAO;
 
@@ -34,7 +41,9 @@ public class OrganizationServiceImpl implements OrganizationService{
 		Organization organization = organizationDAO.read(id);
 		if(organization==null)
 			return null;
-		return organizationDAO.delete(organization);	
-	}
-		
+		if(personDAO.checkPersonInOrganization(organization)!=null)
+			throw new HibernateException("Cannot delete, the Organization still has one or more person belonging to it");
+		else
+			return organizationDAO.delete(organization);
+	}		
 }
